requirements.txt
pillow
colr
opencv-python